fix: let the command-line MCP server start again
- `pnpm dev:mcp` — the local server a developer points Cursor or Claude Desktop at — died on boot: it had to build the whole web server's settings, so it was asked for a port and a public address it does not have just to find out how many cents a search may cost. Settings that belong to one layer moved to that layer, the way the database connection already reads its own url, which also collapsed five copies of the same research-budget block.
- The entry point had drifted behind the toolkit too, missing seven of the services the tools need. A cast on its last line was erasing those unmet requirements, so it type-checked while being unrunnable; dropped it, so the next missing service fails the build instead of puzzling whoever runs it.
- Building the research service used to start its queue workers, which made an editor left connected a second worker on the developer's database — one that took queued runs and ran them, and could spend with real provider keys. A process now says whether it dispatches: the web server does, this surface does not.
- Added BATUDA_ACTIVE_USER_ID, required alongside the organization it already asks for: many tools answer per-person — your mailbox, your tasks — and the ones that write record it as who did it, so a stand-in id would read empty and write rows naming a member who does not exist, which nothing in the database refuses.
