+++ .env.example
@@ -314,3 +314,18 @@ OTEL_EXPORTER_OTLP_ENDPOINT=http://localhost:4318
 # Production example (Honeycomb / Grafana / etc):
 # OTEL_EXPORTER_OTLP_ENDPOINT=https://api.honeycomb.io
 # OTEL_EXPORTER_OTLP_HEADERS=x-honeycomb-team=YOUR_KEY
+
+# ── Local stdio MCP server (`pnpm dev:mcp`) ───────────────────────────
+# Only for the command-line MCP server an editor (Cursor, Claude Desktop)
+# talks to — the web server ignores these. It carries no session, so it has
+# to be told which organization to act in and as whom, and refuses to start
+# without all four. Take the ids from `pnpm cli data users`: the tools that
+# answer per-person read the user id, and the ones that write record it as
+# who did it, so a made-up value reads empty and writes name nobody.
+# BATUDA_ACTIVE_ORG_ID=
+# BATUDA_ACTIVE_ORG_NAME=
+# BATUDA_ACTIVE_ORG_SLUG=
+# BATUDA_ACTIVE_USER_ID=
+# Standing to act with; defaults to the lowest. Set owner or admin to try
+# what someone running the organization can do.
+# BATUDA_ACTIVE_ORG_ROLE=
