+++ .env.example.github
@@ -201,6 +201,10 @@ RESEARCH_CONFIRM_THRESHOLD_FANOUT=10
 
 # mcp-stdio context pinning (apps/server/src/mcp-stdio.ts) — only for the
 # local stdio MCP entrypoint, never set in the production server unikernel.
+# All four are required there. BATUDA_ACTIVE_USER_ID is a seeded member's id
+# (pnpm cli data users): the tools that answer per-person — your mailbox, your
+# tasks — read it, and the ones that write record it as who did it.
 # BATUDA_ACTIVE_ORG_ID=
 # BATUDA_ACTIVE_ORG_NAME=
 # BATUDA_ACTIVE_ORG_SLUG=
+# BATUDA_ACTIVE_USER_ID=
