+++ apps/server/src/main.ts
@@ -1,7 +1,7 @@
 import { createServer } from 'node:http'
 
 import { NodeHttpServer, NodeRuntime } from '@effect/platform-node'
-import { Cause, Config, DateTime, Effect, Layer } from 'effect'
+import { Config, Effect, Layer } from 'effect'
 import {
 	FetchHttpClient,
 	HttpMiddleware,
@@ -9,7 +9,6 @@ import {
 	HttpServerResponse,
 } from 'effect/unstable/http'
 import { HttpApiBuilder, HttpApiScalar, OpenApi } from 'effect/unstable/httpapi'
-import { SqlClient } from 'effect/unstable/sql'
 
 import { BookingProviderLive, IcsParserLive } from '@batuda/calendar'
 import { BatudaApi } from '@batuda/controllers'
@@ -18,7 +17,7 @@ import {
 	ContactDiscovery,
 	makeResearchLlmLive,
 	makeResearchProvidersLive,
-	ResearchEventSink,
+	ResearchDispatch,
 	ResearchService,
 } from '@batuda/research'
 
@@ -53,14 +52,14 @@ import { EnvVars } from './lib/env'
 import { withGlobalMiddlewareOrder } from './lib/global-middleware-order'
 import { LoggerLive } from './lib/logger'
 import { OtlpObservability } from './lib/observability'
+import { ResearchDefaults } from './lib/research-defaults'
 import { WellKnownLive } from './lib/well-known'
 import { McpHttpLive } from './mcp/http'
-import { OrgMiddlewareLive, resolveSystemOrg } from './middleware/org'
+import { OrgMiddlewareLive } from './middleware/org'
 import { SessionMiddlewareLive } from './middleware/session'
 import { ApiKeyService } from './services/api-keys'
 import { CalendarService } from './services/calendar'
 import { CompanyService } from './services/companies'
-import { geocodeCompany } from './services/company-geocoding'
 import { CredentialCrypto } from './services/credential-crypto'
 import { EmailService } from './services/email'
 import { EmailAttachmentStaging } from './services/email-attachment-staging'
@@ -76,16 +75,12 @@ import { OrgResolution } from './services/org-resolution'
 import { PageService } from './services/pages'
 import { PipelineService } from './services/pipeline'
 import { RecordingService } from './services/recordings'
-import { resolveResearchProposedUpdate } from './services/research-apply'
-import { proposalsToAutoApply } from './services/research-auto-apply'
 import { ResearchBlobStorageLive } from './services/research-blob-storage'
+import { ResearchEventSinkLive } from './services/research-event-sink'
 import { ResearchRetention } from './services/research-retention'
 import { S3StorageProviderLive } from './services/s3-storage-provider'
 import { TaskService } from './services/tasks'
-import {
-	ResearchRunCompleted,
-	TimelineActivityService,
-} from './services/timeline-activity'
+import { TimelineActivityService } from './services/timeline-activity'
 import { WebhookService } from './services/webhooks'
 
 const ApiLive = HttpApiBuilder.layer(BatudaApi).pipe(
@@ -114,218 +109,6 @@ const ApiLive = HttpApiBuilder.layer(BatudaApi).pipe(
 		CalendarLive,
 		CalcomWebhookLive,
 	]),
-)
-
-// Wire research event sink → WebhookService + TimelineActivityService
-// Every ending a run reports lands on the timeline so the company activity view
-// surfaces completed research alongside emails and calls — including a run that
-// found nothing usable, which is a result someone still needs to hear. Cost rows
-// stay in research_runs / research_paid_spend.
-// Keyed by the name the sink fires, which is the run's own ending with `run.`
-// swapped for `research.` — so an ending added to TERMINAL_RESEARCH_EVENTS needs
-// a line here too, or runs that end that way leave no trace.
-const TIMELINE_STATUS_FOR_EVENT: Record<
-	string,
-	| 'succeeded'
-	| 'succeeded_low_confidence'
-	| 'failed'
-	| 'cancelled'
-	| 'no_reliable_data'
-	| null
-> = {
-	'research.succeeded': 'succeeded',
-	'research.failed': 'failed',
-	'research.cancelled': 'cancelled',
-	'research.no_reliable_data': 'no_reliable_data',
-}
-
-// ResearchEventSink runs out-of-band of an HTTP request — research runs
-// are kicked off by users but progress events fire later, sometimes from
-// background fibres. Recover the org (and originating user) from the
-// research_run row and enter app_user scope via resolveSystemOrg, so the
-// org-scoped fan-out (webhooks, timeline) writes under RLS like a request
-// instead of relying on the owner connection's bypass.
-const ResearchEventSinkLive = Layer.effect(
-	ResearchEventSink,
-	Effect.gen(function* () {
-		const webhooks = yield* WebhookService
-		const timeline = yield* TimelineActivityService
-		const sql = yield* SqlClient.SqlClient
-		const companyService = yield* CompanyService
-		const geocoder = yield* Geocoder
-		return ResearchEventSink.of({
-			fire: (event, payload) =>
-				Effect.gen(function* () {
-					const researchId = (payload as { researchId?: string }).researchId
-					if (!researchId) {
-						yield* Effect.logWarning(
-							'ResearchEventSink.fire called without researchId in payload',
-						)
-						return
-					}
-					const rows = yield* sql<{
-						organizationId: string
-						createdBy: string | null
-						query: string
-						briefMd: string | null
-						schemaName: string | null
-						status: string
-						paidPolicy: { autoApplyMinConfidence?: number | null } | null
-					}>`
-						SELECT organization_id, created_by, query, brief_md, schema_name,
-							status, paid_policy
-						FROM research_runs
-						WHERE id = ${researchId} LIMIT 1
-					`
-					const [run] = rows
-					if (!run) return
-					// A run that needs reading ends on the same event as any other
-					// success, so the company's own history would record it as clean.
-					// The row itself says which it was, and that is what a person
-					// scrolling back months later has to see.
-					const mapped = TIMELINE_STATUS_FOR_EVENT[event] ?? null
-					const status =
-						mapped === 'succeeded' && run.status === 'succeeded_low_confidence'
-							? 'succeeded_low_confidence'
-							: mapped
-
-					// Fan out as the system actor: load the real org and enter
-					// app_user scope so the timeline write passes RLS like a
-					// request would, instead of leaning on the owner connection's
-					// bypass. webhooks.fire forks its own delivery, so it rides
-					// CurrentOrg (now a real name/slug) but escapes the per-tx GUC.
-					yield* resolveSystemOrg(sql, run.organizationId, {
-						userId: run.createdBy ?? undefined,
-					})(
-						Effect.gen(function* () {
-							yield* webhooks.fire(event, payload)
-							if (!status) return
-							const linkRows = yield* sql<{
-								subjectId: string
-								location: string | null
-								needsCoords: boolean | null
-							}>`
-								SELECT rl.subject_id, c.location, (c.latitude IS NULL) AS needs_coords
-								FROM research_links rl
-								LEFT JOIN companies c
-									ON c.id = rl.subject_id
-									AND c.organization_id = ${run.organizationId}
-								WHERE rl.research_id = ${researchId}
-								  AND rl.subject_table = 'companies'
-								  AND rl.link_kind = 'input'
-								LIMIT 1
-							`
-							const linked = linkRows[0]
-							yield* timeline.record(
-								new ResearchRunCompleted({
-									researchRunId: researchId,
-									companyId: linked?.subjectId ?? null,
-									summary: run.briefMd ?? run.query,
-									status,
-									// The actor is whoever asked for the run, not the background
-									// worker that finished it.
-									actorUserId: run.createdBy ?? null,
-									occurredAt: DateTime.toDateUtc(DateTime.nowUnsafe()),
-								}),
-							)
-
-							// An enrichment run no longer asks the model for coordinates,
-							// so resolve them the deterministic way here: when the run
-							// succeeded for a linked company that has a written location
-							// but no coordinates yet, look the location up in the geocoder
-							// and store lat/long. Best-effort — a miss or failure must
-							// never disturb the timeline and webhook fan-out above.
-							if (
-								status === 'succeeded' &&
-								run.schemaName === 'company_enrichment_v1' &&
-								linked?.location &&
-								linked.needsCoords
-							) {
-								yield* geocodeCompany(linked.subjectId).pipe(
-									Effect.provideService(CompanyService, companyService),
-									Effect.provideService(Geocoder, geocoder),
-									Effect.catchCause(cause =>
-										Effect.logWarning('post-enrichment geocode failed').pipe(
-											Effect.annotateLogs({
-												event: 'research.geocode.failed',
-												companyId: linked.subjectId,
-												cause: Cause.pretty(cause),
-											}),
-										),
-									),
-								)
-							}
-
-							// Confidence-aware auto-apply: when the run's creator set a
-							// threshold, findings that are machine-checkable, verified
-							// deliverable, and confident enough are written to the CRM
-							// without waiting for a person; everything else stays pending
-							// for review. Best-effort per finding — a failure just leaves
-							// that finding pending.
-							// Which suggestions may be written with nobody looking is decided
-							// in one place, so the rule can be checked rather than imitated.
-							// The status comes off the stored row, not the event that ended
-							// the run: a run that needs reading ends on the same event as any
-							// other success.
-							const toApply = yield* proposalsToAutoApply(sql, {
-								researchId,
-								runStatus: run.status,
-								autoApplyMinConfidence:
-									run.paidPolicy?.autoApplyMinConfidence ?? null,
-							})
-							yield* Effect.forEach(toApply, proposedUpdateId =>
-								resolveResearchProposedUpdate(
-									researchId,
-									proposedUpdateId,
-									'apply',
-									null,
-								).pipe(
-									Effect.provideService(CompanyService, companyService),
-									Effect.provideService(Geocoder, geocoder),
-									Effect.provideService(SqlClient.SqlClient, sql),
-									Effect.provideService(TimelineActivityService, timeline),
-									Effect.catchCause(cause =>
-										Effect.logWarning('research auto-apply failed').pipe(
-											Effect.annotateLogs({
-												event: 'research.autoapply.failed',
-												researchId,
-												cause: Cause.pretty(cause),
-											}),
-										),
-									),
-								),
-							)
-						}),
-					).pipe(
-						// Org deleted between run completion and fan-out: skip
-						// rather than crash, mirroring the missing-run return above.
-						Effect.catchTag('SystemOrgNotFound', e =>
-							Effect.logWarning('research fan-out skipped: org not found').pipe(
-								Effect.annotateLogs({
-									event: 'research.fanout.unknown_org',
-									orgId: e.orgId,
-								}),
-							),
-						),
-					)
-				}).pipe(
-					// The event sink is fire-and-forget telemetry. A transient DB
-					// error here must NOT kill the research run, so log and move on;
-					// only a genuine cancellation/shutdown (a pure interrupt) is let
-					// through.
-					Effect.catchCause(cause =>
-						Cause.hasInterruptsOnly(cause)
-							? Effect.interrupt
-							: Effect.logWarning('research event-sink fan-out failed').pipe(
-									Effect.annotateLogs({
-										event: 'research.fanout.failed',
-										cause: Cause.pretty(cause),
-									}),
-								),
-					),
-				),
-		})
-	}),
 )
 
 const ServicesLive = Layer.mergeAll(
@@ -494,7 +277,17 @@ const program = HttpRouter.serve(AppLive, {
 	// brave search, calcom, geocoder, inbox-health-probe, …) can pick up
 	// `HttpClient.HttpClient` from a single shared fetch.
 	Layer.provide(FetchHttpClient.layer),
-	Layer.provide(EnvVars.layer),
+	// Settings, none of which depend on the others, so one provide rather than
+	// three. ResearchDispatch says this is the process that carries queued
+	// research runs out — said here rather than left to its default, because a
+	// default that later changes its mind would stop production running
+	// research with nothing failing: no error, no red test, just runs piling
+	// up queued.
+	Layer.provide([
+		EnvVars.layer,
+		ResearchDefaults.layer,
+		Layer.succeed(ResearchDispatch)(true),
+	]),
 	Layer.provide(PgLive),
 	Layer.provideMerge(ServerLive),
 	// OtlpObservability sits ABOVE LoggerLive so its OTLP logger merges ON TOP of
